+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## x.x.x
 ### PaymentSheet
 * [Fixed] Added a public initializer for `PaymentSheet.BillingDetails`.
+* [Fixed] Fixed some payment method icons not updating to use the latest assets.
+* [Fixed] PaymentSheet no longer displays saved cards that originated from Apple Pay or Google Pay.
 
 ### PaymentsUI
 * [Fixed] And issue with `STPPaymentCardTextField` where the `paymentCardTextFieldDidEndEditing` delegate method was not called.